+++ tests/testthat/tests.make_stancode.R
@@ -881,18 +881,20 @@ test_that("Stan code for non-linear models is correct", {
   flist <- list(a ~ x, b ~ z + (1|g))
   data <- data.frame(
     y = rgamma(9, 1, 1), x = rnorm(9), 
-    z = rnorm(9), g = rep(1:3, 3)
+    z = rnorm(9), v = 1L:9L, g = rep(1:3, 3)
   )
   prior <- c(set_prior("normal(0,5)", nlpar = "a"),
              set_prior("normal(0,1)", nlpar = "b"))
   # syntactic validity is already checked within make_stancode
   scode <- make_stancode(
-    bf(y ~ a - exp(b^z) * (z <= a), flist = flist, nl = TRUE), 
+    bf(y ~ a - exp(b^z) * (z <= a) * v, flist = flist, nl = TRUE), 
     data = data, prior = prior
   )
   expect_match2(scode, 
-    "mu[n] = nlp_a[n] - exp(nlp_b[n] ^ C_1[n]) * (C_1[n] <= nlp_a[n]);"
+    "mu[n] = nlp_a[n] - exp(nlp_b[n] ^ C_1[n]) * (C_1[n] <= nlp_a[n]) * C_2[n];"
   )
+  expect_match2(scode, "vector[N] C_1;")
+  expect_match2(scode, "int C_2[N];")
   
   # non-linear predictor can be computed outside a loop
   scode <- make_stancode(bf(y ~ a - exp(b + z), flist = flist,