+++ tests/testthat/tests.make_standata.R
@@ -303,7 +303,7 @@ test_that("make_standata correctly prepares data for non-linear models", {
   bform <- bf(y ~ a - b^z, flist = flist, nl = TRUE)
   sdata <- make_standata(bform, data = dat)
   expect_equal(names(sdata), 
-    c("N", "Y", "KC", "C", "K_a", "X_a", "Z_1_a_1", 
+    c("N", "Y", "C_1", "K_a", "X_a", "Z_1_a_1", 
       "K_b", "X_b", "Ksp_b", "Imo_b", "Xmo_b_1", "Jmo_b", 
       "con_simo_b_1", "Z_1_b_2", "J_1", "N_1", 
       "M_1", "NC_1", "prior_only")
@@ -311,15 +311,12 @@ test_that("make_standata correctly prepares data for non-linear models", {
   expect_equal(colnames(sdata$X_a), c("Intercept", "x"))
   expect_equal(sdata$J_1, as.array(dat$g))
   
-  sdata <- make_standata(bform, data = dat, control = list(not4stan = TRUE))
-  expect_equal(colnames(sdata$C), "z")
-  
   bform <- bf(y ~ x) + 
     nlf(sigma ~ a1 * exp(-x/(a2 + z))) +
     lf(a1 ~ 1, a2 ~ z + (x|g)) +
     lf(alpha ~ x)
   sdata <- make_standata(bform, dat, family = skew_normal())
-  sdata_names <- c("C_sigma", "X_a2", "Z_1_a2_1")
+  sdata_names <- c("C_sigma_1", "X_a2", "Z_1_a2_1")
   expect_true(all(sdata_names %in% names(sdata)))
 })